move Diff from test to main
