+++ lib/src/test/java/com/ktown4u/utils/DiffTest.java
@@ -1,13 +1,8 @@
 package com.ktown4u.utils;
 
-import com.github.difflib.text.DiffRow;
-import com.github.difflib.text.DiffRowGenerator;
 import org.approvaltests.Approvals;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.Arrays;
-import java.util.List;
 
 public class DiffTest {
     @Test
@@ -71,39 +66,4 @@ public class DiffTest {
                 """;
     }
 
-    private class Diff {
-        private final String before;
-        private final String after;
-
-        public Diff(final String before, final String after) {
-            this.before = before;
-            this.after = after;
-        }
-
-        public String print() {
-            final DiffRowGenerator generator = DiffRowGenerator.create()
-                    .showInlineDiffs(true)
-                    .mergeOriginalRevised(true)
-                    .inlineDiffByWord(true)
-                    .oldTag(f -> f ? "" : " -> ")
-                    .newTag(f -> "")
-                    .build();
-
-            final List<DiffRow> rows = generator.generateDiffRows(
-                    Arrays.asList(before.split("\n")),
-                    Arrays.asList(after.split("\n"))
-            );
-
-            String result = "";
-
-            for (final DiffRow row : rows) {
-                if (DiffRow.Tag.EQUAL == row.getTag()) result += row.getOldLine() + "\n";
-                if (DiffRow.Tag.CHANGE == row.getTag()) result += "+ " + row.getOldLine() + "\n";
-                if (DiffRow.Tag.DELETE == row.getTag()) result += "-- " + row.getOldLine() + "\n";
-                if (DiffRow.Tag.INSERT == row.getTag()) result += "++ " + row.getNewLine() + "\n";
-            }
-
-            return result;
-        }
-    }
 }
